requirements.txt
requests
fastapi
SQLAlchemy
uvicorn
starlette==0.19.1
pydantic
setuptools==65.5.1
python-dotenv
ffmpeg-python==0.2.0
psycopg2
pydantic[email]
python-multipart